Adds command to enumerate volumes
The primary purpose of this is to be able to re-use old volumes which
are no longer referenced by any volumes, but also to build on it later.

To facitiltate re-using old volumes "--volumes-from" has been modified
to optionally accept a volume id and a mount path.

With this change there is no longer any need for data-containers at all and one
of the use-cases for bind-mounts.

This change has no API breakage and only adds as minimal an API required
for the above use-case.

Docker-DCO-1.1-Signed-off-by: Brian Goff <cpuguy83@gmail.com> (github: cpuguy83)