POC Adds introspection socket injection support
Inject introspection socket into container using normal docker http
server.

Modifies api/server.go to have ListenAndServe(now renamed) to return a
net.Listener, this enables us to track the started servers for any
container using an introspection socket and subsequently close it when
the container is stopped.

If/when libchan support is added, it just needs to be added as a
supported api/server type (e.g. chan:// instead of tcp:// or unix://).