push/pull attach stdout/stderr of other pull/push
Fixes #8385

When there is already a push/pull in progress for an image, instead of
just reporting "another client is already ...", attach to the
stdout/stderr of the other job so push/pull progress is reported to new
client.

For example:

```bash
docker pull debian:jessie &
docker pull debian:jessie
```

Both clients will have the pull progress reported instead of just the
first.
The example is contrived, but illustrates when two clients are trying to
pull the same image.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>